vault helpers for the UI

+++ contracts/interfaces/ILockupHell.sol
@@ -4,6 +4,6 @@ pragma solidity >=0.8.0;
 
 interface ILockupHell {
     function lockupRewards(address user, uint256 shortLockup, uint256 longLockup) external;
-    function getShortPercentage() external view returns(uint32); 
-    function getLongPercentage() external view returns(uint32);
+    function getShortPercentage() external view returns(uint256); 
+    function getLongPercentage() external view returns(uint256);
 }
